requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gcloud==0.18.3
googleapis-common-protos==1.66.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.5
jwcrypto==1.5.6
MarkupSafe==3.0.2
oauth2client==4.1.3
packaging==24.2
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pyparsing==3.2.1
Pyrebase4==4.8.0
python-decouple==3.8
python-jwt==4.1.0
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
